+++ src/analyst/autoencoder_feature_generator.py
@@ -1464,7 +1464,7 @@ class AutoencoderFeatureGenerator:
                 self.logger.warning(f'⚠️ Trial failed: {str(e)}')
                 return float('inf')
         # Use new Bayesian TPE optimizer
-        from src.utils.ml_common.optimization.bayesian_tpe_optimizer import (
+        from src.utils.nas_tas.bayesian_tpe_optimizer import (
             BayesianTPEOptimizer,
             BayesianTPEConfig
         )
