+++ src/analyst/sr_relevance_optimizer.py
@@ -65,7 +65,7 @@ class SRRelevanceOptimizer:
             performance = self._evaluate_weights(weights, historical_data, detected_sr_levels, actual_outcomes)
             return -performance
         # Use new Bayesian TPE optimizer
-        from src.utils.ml_common.optimization.bayesian_tpe_optimizer import (
+        from src.utils.nas_tas.bayesian_tpe_optimizer import (
             BayesianTPEOptimizer,
             BayesianTPEConfig
         )
