+++ src/research/profit_labeling/dynamic_target_optimizer.py
@@ -802,7 +802,7 @@ class JointTargetHorizonOptimizer:
                            val_data: pd.DataFrame) -> OptimizationResult:
         """Perform Bayesian TPE optimization using unified optimizer."""
         # Import Bayesian TPE optimizer
-        from src.utils.ml_common.optimization.bayesian_tpe_optimizer import (
+        from src.utils.nas_tas.bayesian_tpe_optimizer import (
             BayesianTPEOptimizer,
             BayesianTPEConfig
         )
