+++ src/research/profit_labeling/parameter_optimizer.py
@@ -484,7 +484,7 @@ class ParameterOptimizer:
                            val_data: pd.DataFrame) -> OptimizationResult:
         """Bayesian TPE optimization using unified optimizer."""
         # Import Bayesian TPE optimizer
-        from src.utils.ml_common.optimization.bayesian_tpe_optimizer import (
+        from src.utils.nas_tas.bayesian_tpe_optimizer import (
             BayesianTPEOptimizer,
             BayesianTPEConfig
         )
