+++ src/training/steps/backtesting/final_parameters_optimization.py
@@ -24,7 +24,7 @@ import numpy as np
 import logging
 
 # Import new Bayesian TPE optimizer
-from src.utils.ml_common.optimization.bayesian_tpe_optimizer import (
+from src.utils.nas_tas.bayesian_tpe_optimizer import (
     BayesianTPEOptimizer,
     BayesianTPEConfig,
     optimize_with_bayesian_tpe
