+++ src/training/steps/backtesting/real_parameters_optimization.py
@@ -49,7 +49,7 @@ except ImportError:
     optuna = None
 
 # Import new Bayesian TPE optimizer
-from src.utils.ml_common.optimization.bayesian_tpe_optimizer import (
+from src.utils.nas_tas.bayesian_tpe_optimizer import (
     BayesianTPEOptimizer,
     BayesianTPEConfig,
     optimize_with_bayesian_tpe
