+++ src/training/steps/market_analysis/feature_lookback_optimization/mrmr_lookback_optimizer.py
@@ -42,7 +42,7 @@ except ImportError:
     logging.warning("Optuna not available - using fallback optimization")
 
 # Import new Bayesian TPE optimizer
-from src.utils.ml_common.optimization.bayesian_tpe_optimizer import (
+from src.utils.nas_tas.bayesian_tpe_optimizer import (
     BayesianTPEOptimizer,
     BayesianTPEConfig,
     optimize_with_bayesian_tpe
