+++ src/training/steps/market_analysis/hybrid_nas_tas_regime/shared_utils/unified_search_engine.py
@@ -66,7 +66,7 @@ from src.utils.tprint import (
 
 # Import ML common utilities
 try:
-    from src.utils.ml_common.optimization.bayesian_tpe_optimizer import (
+    from src.utils.nas_tas.bayesian_tpe_optimizer import (
         BayesianTPEOptimizer, BayesianTPEConfig, OptimizationResult
     )
     ML_COMMON_AVAILABLE = True
