+++ src/training/steps/market_analysis/nas_modeling/core/nas_trainer.py
@@ -118,7 +118,7 @@ except ImportError:
 
 # Import ML optimization utilities
 try:
-    from src.utils.ml_common.optimization.bayesian_tpe_optimizer import (
+    from src.utils.nas_tas.bayesian_tpe_optimizer import (
         BayesianTPEOptimizer, BayesianTPEConfig, OptimizationResult,
         optimize_with_bayesian_tpe, create_search_space_from_bounds
     )
