+++ src/training/steps/market_analysis/optimized_multi_horizon_optimizer/grid_bayesian_optimizer.py
@@ -24,7 +24,7 @@ from src.utils.ml_common.optimization.grid_utils import (
 )
 from src.utils.ml_common.utils.hpo_utils import HyperparameterOptimization
 # Import new Bayesian TPE optimizer
-from src.utils.ml_common.optimization.bayesian_tpe_optimizer import (
+from src.utils.nas_tas.bayesian_tpe_optimizer import (
     BayesianTPEOptimizer,
     BayesianTPEConfig,
     optimize_with_bayesian_tpe
