+++ src/training/steps/market_analysis/tas_regime/evaluation/tree_evaluator.py
@@ -698,7 +698,7 @@ class AdvancedTreeEvaluator:
         
         try:
             # Import Bayesian TPE optimizer
-            from src.utils.ml_common.optimization.bayesian_tpe_optimizer import (
+            from src.utils.nas_tas.bayesian_tpe_optimizer import (
                 BayesianTPEOptimizer,
                 BayesianTPEConfig
             )
