+++ src/training/steps/market_analysis/tas_regime/shared_utils/search_strategies.py
@@ -595,7 +595,7 @@ class SearchStrategies:
         """Manual Bayesian optimization implementation using unified Bayesian TPE optimizer."""
         try:
             # Import Bayesian TPE optimizer
-            from src.utils.ml_common.optimization.bayesian_tpe_optimizer import (
+            from src.utils.nas_tas.bayesian_tpe_optimizer import (
                 BayesianTPEOptimizer,
                 BayesianTPEConfig
             )
