+++ src/training/steps/model_training/bayesian_optimization_msm.py
@@ -33,7 +33,7 @@ except ImportError:
     OPTUNA_AVAILABLE = False
 
 # Import new Bayesian TPE optimizer
-from src.utils.ml_common.optimization.bayesian_tpe_optimizer import (
+from src.utils.nas_tas.bayesian_tpe_optimizer import (
     BayesianTPEOptimizer,
     BayesianTPEConfig,
     optimize_with_bayesian_tpe
