+++ src/training/steps/model_training/enhanced_regime_aware_hpo.py
@@ -49,7 +49,7 @@ except ImportError:
     OPTUNA_AVAILABLE = False
 
 # Import new Bayesian TPE optimizer
-from src.utils.ml_common.optimization.bayesian_tpe_optimizer import (
+from src.utils.nas_tas.bayesian_tpe_optimizer import (
     BayesianTPEOptimizer,
     BayesianTPEConfig,
     optimize_with_bayesian_tpe
