+++ src/utils/ml_common/models/enhanced_model_trainer.py
@@ -414,7 +414,7 @@ class EnhancedModelTrainer:
         """Perform hyperparameter optimization using unified Bayesian TPE optimizer."""
         try:
             # Import Bayesian TPE optimizer
-            from src.utils.ml_common.optimization.bayesian_tpe_optimizer import (
+            from src.utils.nas_tas.bayesian_tpe_optimizer import (
                 BayesianTPEOptimizer,
                 BayesianTPEConfig
             )
