+++ src/utils/ml_common/nas_tas_unified/search.py
@@ -20,7 +20,7 @@ from abc import ABC, abstractmethod
 
 # Import Bayesian TPE optimizer
 try:
-    from src.utils.ml_common.optimization.bayesian_tpe_optimizer import (
+    from src.utils.nas_tas.bayesian_tpe_optimizer import (
         BayesianTPEOptimizer, BayesianTPEConfig, optimize_with_bayesian_tpe
     )
     BAYESIAN_TPE_AVAILABLE = True
