+++ src/utils/ml_common/optimization/bayesian_entry_timing_optimizer.py
@@ -26,7 +26,7 @@ except ImportError:
     warnings.warn("Optuna not available. Install with: pip install optuna")
 
 # Import new Bayesian TPE optimizer
-from src.utils.ml_common.optimization.bayesian_tpe_optimizer import (
+from src.utils.nas_tas.bayesian_tpe_optimizer import (
     BayesianTPEOptimizer,
     BayesianTPEConfig,
     optimize_with_bayesian_tpe
