+++ src/utils/ml_common/optimization/hierarchical_hpo.py
@@ -335,7 +335,7 @@ class HierarchicalHPO:
                 )
             
             # Configure Bayesian TPE optimizer
-            from src.utils.ml_common.optimization.bayesian_tpe_optimizer import (
+            from src.utils.nas_tas.bayesian_tpe_optimizer import (
                 BayesianTPEOptimizer,
                 BayesianTPEConfig
             )
