+++ src/utils/ml_common/optimization/hpo_utils.py
@@ -712,7 +712,7 @@ class HyperparameterOptimization:
                 self.logger.info("🚀 Using enhanced non-linear search space")
 
             # Import Bayesian TPE optimizer
-            from src.utils.ml_common.optimization.bayesian_tpe_optimizer import (
+            from src.utils.nas_tas.bayesian_tpe_optimizer import (
                 BayesianTPEOptimizer,
                 BayesianTPEConfig
             )
