+++ src/utils/ml_common/optimization/neural_architecture_search.py
@@ -61,7 +61,7 @@ except ImportError:
     OPTUNA_AVAILABLE = False
 
 # Import new Bayesian TPE optimizer
-from src.utils.ml_common.optimization.bayesian_tpe_optimizer import (
+from src.utils.nas_tas.bayesian_tpe_optimizer import (
     BayesianTPEOptimizer,
     BayesianTPEConfig,
     optimize_with_bayesian_tpe
